# labscript state machine modification and optimization