## Background
This is the first set of changes to optimize the runtime of labscript† for short time-scales (20-100ms). labscript seemed particularly sub-optimal in the use-case of executing a large sequence of queued up shots. This PR addresses the following bottlenecks:

1. Calling `transition_to_manual` between every shot
    - when shots are queued up back to back, the user is not using BLACS in manual mode
2. All functionality is executed on the QT main thread
3. All workers of a device are processed serially 
4. Master Pseudoclock `status_monitor` scheduled every 100ms

## Changes
1. Break down `transition_to_manual` state worker task, do not call when shots queued up
    - Introduce new `post_experiment` state that will only run the experiment post processing/saving
    - `transition_to_manual` manual handles the remaining tasks that involve setting up manual mode operation of the device
    - as a result, `transition_to_manual` is only ever called when it is possible for the user to manually interface with the hardware (i.e. there are no shots in the queue, the user pauses the queue, a failure has occured)
    - Backwards compatibility added to support devices that don't have `post_experiment` state implemented
2. Remove inmain decorators, allow execution from `DeviceTab.mainloop`, `ExperimentQueue.manager`, etc.
    - State GUI functions no longer run in main - all functions that make updates to the GUI to grab the Qtlock before making the update
    - Make StateQueue implementation thread safe using local locks
3. Parallelize the workers in each device
    - Send all tasks in a list to the generator function
    - `DeviceTab.mainloop` starts all workers at once and waits to hear back from them
4. Schedule Pulseblaster status_monitor checks a more frequently
    - schedule status_monitor every 1ms

## Results
1. Allows us to save time by not making expensive calls to program devices to manual mode 
    - Making the NI DAQmx driver calls for the NI-6363 acquisition worker is particularly time-consuming. This change saves us **~50ms**.
    - Note, this is not the case with most devices - if their manual mode programming is already short, they will not be affected by this change. 
2. Eliminates unnecessary delays in the scheduling of tasks
    - By executing all functions from the thread they are called in, we are minimizing the latency of when tasks are started and maximizing the concurrency of the program. This saves us **~25-30ms**
    - With local locking, there are no issues/race-conditions with updating the GUI
3. Straightforward parallelization of the worker processes 
    - The NI-6363 is a multifunction card with multiple workers. The output worker takes ~50ms and acquisition worker takes ~75ms. This change saves us **~50ms**
    - Note, most devices don't have multiple workers. They will not be affected by this change
4. Allows us to know the experiment is done as soon as possible
   - Saves **~15-25ms**

Overall, these changes bring down the overhead between shots in our experimental set up from **380ms to 220ms**.

†_Some of the optimizations are specific to my experimental set up and the hardware/functionalities I am testing with. You may not see the same level of improvements I achieved when applying to your experiment._